[API] Add IHttpService and implementation
To achieve maximal testibility the communication with the web is
abstracted out into an HttpService. It has a custom returntype,
IHttpRespone, wich is implemented in HttpResponse. A new test to make sure
this implementation converts the System.Net.HttpStatusCode into the
correct int and string representations gets added to API.Tests.